core: io stub gains Writer + Copy; Reader/Writer return nint
The hand-finished io stub predates the Go-int -> nint mapping: its
Reader returned (int n, error err), which no converted implementation
could satisfy through the generated forwarding (tuple conversions are
not element-wise). Reader/Writer now return nint, and the stub gains
the Writer interface and a faithful Copy - the surface the new
AnonymousInterfaces guard (archive/tar's ReadFrom-hiding shape) needs.

+++ src/core/io/io_Reader.cs
@@ -86,7 +86,7 @@ public static partial class io_package
         private static readonly ReadByVal? s_ReadByVal;
 
         [DebuggerNonUserCode, MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public (int, error) Read(in slice<byte> p)
+        public (nint, error) Read(in slice<byte> p)
         {
             T target = m_target;
